Fix 'bytes2rawstr()' for non-Unicode chars
Use the 'backslashreplace' with 'bytes.decode()' so that the bytes
decoding into a raw string doesn't get aborted in the presence on
invalid Unicode chars. It will now decode those chars with '\xNN'
instead of aborting the decoding.

Required because for scoring English strings, we need decrypted byte
arrays to be translated into strings. Although wee could simply return
an empty string when catching the 'UnicodeDecodeError' exception, it's
preferable to keep on trying to decode more chars and just mark the ones
that cannot be decoded as invalid.