Rework C3 implementation to be reusable
Make C3 implementation to also return the score associated with the
decrypted message.

This way, this challenge functions may be reused in case we need to get
decrypted messages and scores for different encrypted messages, so that
the scores can be compared against each other, which is the case of C4.