Implement Challenge 5: Implement repeating-key XOR
Add script for implementing Challenge 5.

Use same script structure as in previous challenges implementation.
The challenge input data was placed in a new data file 'data_c5.txt'.
After reading the data file, a check must be made for removing a
possible trailing '\n' from the string representing the file data that
was read. The trailing '\n' is (correctly) added by some editors at the
last line of a text file, because POSIX forces all lines to terminate
with a newline char.

Using a file as entry point for getting the encrypted data (instead
of a string) because the challenge requests to exercise it on several
local files as well:

 "...
  Encrypt a bunch of stuff using your repeating-key XOR function.
  Encrypt your mail. Encrypt your password file. Your .sig file. Get
  a feel for it. I promise, we aren't wasting your time with this.
 ..."

Reuse functionality from 'cp_test_c02', because this challenge is
basically about XOR'ing data. Add an extra param to C2's 'execute_xor()'
function, to specify if the 'src' and 'key' strings are hex encoded or
raw strings, adapting the function's implementation accordingly. C2 used
hex encoded strings as input but this challenge uses raw strings.

Move the 'file_get_lines()' function to the 'utils' module because it's
a generic routine. Add a new 'file_get()' function for reading a whole
file in, also placed in the 'utils' module and required for this
challenge implementation.