Implement Challenge 7: AES in ECB mode
Add script for implementing Challenge 7.

Use same script structure as in previous challenges implementation.
The challenge input data was placed in a new data file 'data_c7.txt'.

The expected results are not declared in the challenge's statement.
They came out from test execution and debugging.
The results seem correct, as they turned out to be the same as in
challenge 6 (followed by 4 extra 0x04 bytes due to padding maybe?). This
way, used 'data_c6_out.txt' for checking the results.

The 'get_ciphertext()' function from C6 was reused in this challenge for
getting the ciphertext, which is base64 encoded in a text file.

The challenge depends on the ''Crypto.Util.AES' module for accessing the
AES functionality required by this challenge.