Implement Challenge 8: AES in ECB mode
Add script for implementing Challenge 8.

Use same script structure as in previous challenges implementation.
The challenge input data was placed in a new data file 'data_c8.txt'.

The expected results are not declared in the challenge's statement.
They came out from test execution and debugging.
Not sure about the results correctness, but for now assuming they are
correct, so set the test's expected data accordingly.