Move AES decrypt function to utilities module
Move the implementation of the AES decrypt function from challenge 7
into the utilities module.

The function will be useful for other challenges.
Additionally, allows keeping the challenges depending only on the utils
module and hide the actual implementation details and dependencies (in
this case from the 'Crypto.Cipher.AES' module) inside the utilities
module.