Implement Challenge 12: Simple Byte ECB decryption
Add script for implementing Challenge 12.

Use same script structure as in previous challenges implementation.

The expected results are not explicitly declared in the challenge's
statement, although they can easily be seen by base64 decoding the
string from the challenge's statement. This is exactly what was done for
setting the challenge's test expected result.

To note that an extra '\x01' byte needed to be added to the expected
result in order for it to match the obtained results. This is due to
PKCS#7 padding done inside the encryption oracle function.