Implement Challenge 14: Harder Byte ECB decryption
Add script for implementing Challenge 14.

Use same script structure as in previous challenges implementation.

Based on Challenge 12 implementation, which is basically a particular
case of Challenge 14, where the prefix added by the oracle has zero
length. This way, this challenge imports the C12 implementation for
reusing some of its functions.

The expected result is also the same as in C12, so this script reuses
the results file 'data_c12_out.txt' for confirming the test.