Merge C12 and C14 implementations in C14
Challenge 12 is exactly the same as Challenge 14 with a zero sized
prefix (i.e., no prefix used).

Reworked C14 so that the oracle can be configured to not add a prefix
(to serve C12's purpose). This was done by adding a flag parameter to
the 'execute_break_ecb()' function that tells if the oracle is to add a
prefix or not. For executing C14, this flag is 'True'.

All functions from C12 that were being used by C14 were moved into C14,
so now C14 does not depend anymore on C12. Instead, C12 is now importing
from C14 and just calling C14's 'execute_break_ecb()' function with the
"use prefix" flag set to 'False', so that the oracle doesn't use any
prefix.