Implement Challenge 15: PKCS#7 padding validation
Add script for implementing Challenge 15.

Use slightly changed script structure compared to previous challenges
implementation with respect to test results validation, as described
below.

The script still calls its 'execute' function, which in turn calls the
utilities PKCS#7 unpad function. The unpad functions is already throwing
exceptions when the padding is not valid (recommended in the challenge's
statement).

The validation of the results is done by making sure that exceptions are
caught for the test strings that are incorrectly padded and that no
exception is thrown when the test strings are correctly padded.