Implement Challenge 17: The CBC padding oracle
Add script for implementing Challenge 17.

Use same script structure as in previous challenges implementation, with
some additional functions as required and described by the challenge's
statement.

Besides asking the encryption oracle to select a random string from the
challenge's string set ('pt_strings' in the code), which is done by
passing '-1' to the oracle's 'string_idx' parameter, it's also possible
to request it to use a specific string, providing a valid 'pt_strings'
index (0 based) in 'string_idx'.

The challenge's test script is using specific string indexes and doing a
round for each of the strings in the set, so that it can prove the
breaking of all ciphertexts using the CBC padding oracle attack, which
goes a bit beyond the challenge's statement. The script can easily be
changed to mimic exactly the statement's behavior, by replacing the loop
that calls 'execute_cbc_padding_oracle_attack()' for each string with a
single call with '-1' as parameter.

The results validation is done by comparing the broken returned
plaintext against the original string before encrypting it.