Implement Challenge 18: Implement CTR
Add script for implementing Challenge 18.

Use same script structure as in previous challenges implementation, with
2 "execute()" functions, one for encrypting and another for decrypting.

The results validation is done by confirming that decrypting and then
re-encrypting the challenge's statement input string produces the exact
same string.

Utility module was extended with a class ('AesCtr') for implemented the
CTR functions, currently encrypt and decrypt. A class was chosen for
simplicity of use, but mostly because the CTR implementation needs to
maintain state between calls to encrypt/decrypt.