Implement Challenge 19: Break fixed-nonce CTR mode
Add script for implementing Challenge 19, for breaking fixed-nonce CTR
mode using substitution.

Challenge input strings were placed in the 'data_c19.txt' file.
Result validation is done by comparing the broken plaintext strings
against the challenge input ones.

Two different methods were implemented for breaking the CTR plaintexts,
both based on the fact that with CTR using the same fixed-nonce, the XOR
of two ciphertexts is the XOR of the two associated plaintexts:

 * 'guess_key_1()':

    This method's strategy was based on brute forcing each key position.
    For each key byte, all values are tried. For each value we assess
    the resulting decrypted byte at the same position for all
    ciphertexts we have (that have text in that position - some are
    shorter), by building a string with all those bytes and score it for
    English (letter frequency, etc.).
    For example for the 1st key byte, we try all 256 values and for each
    value we take all the 1st decrypted bytes on all ciphertexts, build
    a string with them and score it. The key byte value giving the
    highest score is assumed to be the correct key byte.
    This is much like "Challenge 6: Break repeating-key XOR".

 * 'guess_key_2()':

    This is a "crib-dragging" method using the space as a crib. Since a
    space XORed with any letter will give the same letter with its case
    changed, the method XORs ciphertexts with each other looking for
    letters, which will indicate that there may be a space in the
    respective position on the plaintexts. Knowing which will allow
    deriving the key stream bytes at those space positions.
    This method is thoroughly described in the web. Search for "crib
    dragging" and "Many time pad attacks".

After having implemented and tested both methods described above, it was
noticed that the best results were obtained by combining both methods,
but giving priority to the first one. Considering the goal is to get as
many key stream bytes cracked as possible, the best method was to take
all the key bytes cracked by method 1 and get any missing key bytes from
the ones given by method 2, if available.

Even by combining both methods, some off the key stream bytes could not
be cracked. This happens because some plaintexts are longer than other
so it becomes harder and more error prone to crack those key bytes on
the end off the longer plaintexts. Although it could still be possible
to apply somehow both methods above on the longer ciphertexts, by
looking for dictionary words, scoring the plaintexts, etc., it was
concluded that a manual approach to crack the remaining bytes was faster
and with good results. This is also described in the literature related
to these attacks.

The manual guessing method described above is implemented in the
function 'guess_key_manually()'. It shows the user the current key bytes
that have been cracked and the cracked plaintexts results. This will
look almost like the real plaintexts, with just some small parts missing
or some obvious errors (like incorrect case, etc.).
The user can then change the errors or try guessing any missing
plaintext parts and the function will, with that info, change or set the
associated key bytes resulting in the recalculation and display of all
the plaintexts again given the new key. The function also indicates if
the current cracked plaintexts are matching the real ones or not.

Since after the 2 methods above, only a few parts of plaintexts will be
missing (near each plaintext end), by checking the plaintexts context
and comparing plaintexts against each other, it becomes relatively easy
to crack the missing/incorrect parts using the manual guessing process
interactively in a few rounds.

Other functions used:

 * 'debug_msg()' and 'dump_list()': For test/debug purposes.

 * 'ctr_encrypt_string_list()' and 'ctr_encrypt()': Do the encryption of
   the plaintexts.

 * 'ctr_decrypt_string_list()': Decrypts the plaintexts using the
   current cracked key.

 * 'verify_break()': Confirms if the plaintexts currently broken are
   matching the real ones.

 * 'execute_break_ctr_fixed_nonce()': The challenges main control
   function. Calls the automatic key break methods and then controls the
   manual key guessing process. Finally, it validates if the plaintexts
   were successfully broken or not.