Implement Challenge 20: Break fixed-nonce CTR mode
Add script for implementing Challenge 20, for breaking fixed-nonce CTR
mode statistically.

Challenge input strings were placed in the 'data_c20.txt' file.
Result validation is done by comparing the broken plaintext strings
against the challenge input ones.

Since the goal is basically the same as for Challenge 19, this challenge
is using C19 functionality directly.

For implementing exactly the process described in the challenge's
statement, a new guess method was added to C19, implemented via the
function 'guess_key_3()', which in turn is based on Challenge 6
'break_key()' function for breaking a repeating-key XOR ciphertext.

Using of the 'guess_key_3()' has been tested, but it turned out that the
'guess_key_1()' function used for C19 already included the same results
as 'guess_key_3()'. Furthermore, it can even guess more key bytes,
because it's not limited to the smaller plaintext, i.e., it continues to
do the same process for more key bytes, even if using lesser
ciphertexts for increasing key positions, because some of them end at
lower positions.

This way, function 'guess_key_3()' was left in the code for reference
only. It's not being used for neither C19 nor in C20.
In fact, C20 is exactly like C19 with a different plaintext string set
as input.