Implement Challenge 23: Clone an MT19937 RNG
Add script for implementing Challenge 23.

Use same script structure as in previous challenges implementation.

Function 'mt19937_untemper()' is the "untemper" function mentioned in
the challenge's statement. It's built on top of utility functions
'undo_rs_xor()' and 'undo_ls_and_xor()', which abstract the Right Shift
-> XOR and Left Shift -> AND -> XOR operations done inside the MT19937
RNG for providing an output number which results from tempering its
internal state. The "untemper" function is basically the execution, in
reversed order, of the inverse functions (the 'undo' functions) used in
tempering.

The test runs like this from its main entry point:

 * The Real RNG is created via 'mt19937_create()'. It uses a random seed
   and, immediately after creating, a random amount of numbers are
   extracted from it, without being saved. The purpose is to highlight
   the fact that cloning is possible no matter what the state of the
   Real RNG is.

 * The 'execute_mt19937_clone()' is called for doing the actual cloning,
   provided with the Real RNG (required for output capturing) and a
   count for the amount of numbers, generated from both the clone and
   the real RNG, that the function should return in the end for
   validation.
   The cloning itself is executed as described in the statement: A new
   RNG (the clone) is created with some seed (the seed is irrelevant
   because we will replace its full state anyway); The real RNG is used
   for capturing the 624 outputs, untemper them and build the clone's
   internal state from the untempered outputs; Once the internal state
   is fully cloned, the clone RNG internal state is overwritten with the
   cloned state, using the new 'set_state()' method from the 'MT19937'
   class, added for this purpose; Finally, a list of 'count' numbers is
   extracted from each RNG (real and clone) and returned for validation.

 * Finally, the main script checks if the list of numbers generated from
   the Real RNG matches the on generated from the Cloned RNG. The test
   succeeds when the lists are a perfect match. When failing, the script
   provides some info about the failure (in which number the difference
   was found and the respective numbers on both lists).

The utility class 'MT19937' (for abstracting the RNG) was extended with
a new 'set_state()' method, which allows replacing the RNG's current
internal state array with the one provided via the function. This is the
method used for splicing the cloned state into the clone RNG instance.