Implement Challenge 26: CTR bitflipping
Add script for implementing Challenge 26.

This challenge is a re-edition of "Challenge 16: CBC bitflipping
attacks.", but using AES-CTR instead of AES-CBC, so the structure and
basic principles are the same.

In fact, with CTR it's even simpler to perform the attack, because
there's no padding and block alignment requirements involved (stream
cipher property, unlike block ciphers) and because for changing bits in
the plaintext, we change the same bits on the associated ciphertext,
unlike CBC where the bits we flip on a block, affect the plaintext in
the next block.

The functions are the same as in C16, with small adaptations:

 * 'escape_userdata()': For escaping the '=' and ';' chars.
   Exactly the same as in C16.

 * 'encryption_oracle()': The oracle of encryption.
   Same principle as in C16, but using CTR for the encryption instead of
   CBC.

 * 'decryption_oracle()': The decryption oracle.
   Same principle as in C16, but using CTR for the decryption instead of
   CBC.

 * 'execute_ctr_bitflip_attack()': The attack core implementation.
   Uses same principle as in C16, but there's only one piece of
   ciphertext and associated plaintext involved (no multiple blocks).

The main script simply calls 'execute_ctr_bitflip_attack()', presents
the results and assesses the test success.