Implement Challenge 28: Implement SHA-1 keyed MAC
Add script for implementing Challenge 28.

Following the usual test script structure.
Added new utility class 'SHA1' and function 'hmac_sha1()' to module
'cp_aux_utils', for performing the required SHA-1 hashing operations.

The test script performs SHA-1 keyed hashing using a random sized (1 to
32 bytes) key with random bytes in it. This is implemented in function
'execute_test_sha1()'.

For confirming that SHA-1 hashing is correct and also for trying to get
a SHA-1 hash without knowing the key, the 'execute_test_sha1()' function
also provides the message hashed without a key. To confirm that the hash
algorithm is correct, the non-keyed hash This is compared to the hash
produced on the same message using 'sha1sum' linux tool.

The test script, as required in the challenge's statement, also does a
simple verification that a tampered message will result in a different
hash being produced and at the same time that we cannot produce a valid
hash on the tampered message without knowing the key.