Implement Challenge 29: Break a SHA-1 keyed MAC
Add script for implementing "Challenge 29: Break a SHA-1 keyed MAC using
length extension". Following the usual test script structure.

Added new utility function 'sha1_pad()' for performing padding as
required for SHA-1 and needed by this challenge.

Reworked the 'SHA1' class constructor so that a 'SHA1' object can be
created with a specified and specific initial state, besides the default
initial state. This is also required by the challenge.

The core challenge function is 'execute_break_sha1()', which:

 * Receives the original message and the string for appending, as
   described in the challenge's statement.

 * Starts by getting the hash for the original message. The hashing
   function uses a key with size up to 'max_key_sz' (1024 bytes).

 * For each possible key size (from 1 to 'max_key_sz') requests
   'sha1_extend()' to provide the forged message and associated hash and
   then checks if the returned hash matches the hash provided by the
   real hashing function. If it does, then the key size is right which
   in turn means the associated forged message and hash are also correct
   and the goal is achieved.

The 'sha1_extend()' function implements the core breaking process for a
given key size, as described in the challenge's statement:

 * Based on a tentative key size, determines the padding that is used by
   the hashing function on the original message and uses it to build the
   tentative forged message (original + padding + append).

 * Uses the original message hash and tentative forged message length to
   clone the SHA1 state left by the real hash function at the end of the
   original message hashing. This in fact creates a new SHA1 hasher
   whose initial state matches the final state of the hasher used on the
   original message.

 * Anything fed to the cloned hasher at this point will produce exactly
   the same hash as if it was fed to the original hasher after the
   padded original message, thus being a valid hash for the secret key.

 * Feed the cloned SHA1 hasher with the string we want to append.

 * Return the forged message candidate and respective hash so that
   'execute_break_sha1()' can validate if they are the correct ones
   (they will when the key size matches the real one).

The main script simply calls 'execute_break_sha1()' for collecting the
forged message and associated hash, which will be 'None' if the execute
function could not find them. The execute function also returns the
broken key size. The collected data is then presented and used for
determining the test success (forged message and hash are not 'None' and
the broken key size matches the real one).