Implement Challenge 30: Break a MD4 keyed MAC
Add script for implementing "Challenge 30: Break an MD4 keyed MAC using
length extension".

This challenge is basically a copy of Challenge 29 that uses MD4 for
hashing instead of SHA-1. The basis for the attack is exactly the same,
so refer to the Challenge 29 commit (65392ce) for all the technical
details.

New utilities were added for dealing with MD4:

 * New class 'MD4' for the core MD4 hashing implementation. Besides
   using the default initial state in MD4, it's also possible to create
   the object with some predefined initial state (required by the
   breaking process).

 * New function 'md4_mac()' for producing a keyed MAC using MD4.

 * New function 'md4_pad()' for performing MD4 padding.

The script structure, presented results and test success determination
is also exactly the same as for Challenge 29.