Implement Challenge 31: Break HMAC-SHA1
Add server and client scripts for implementing "Challenge 31: Implement
and break HMAC-SHA1 with an artificial timing leak".

This challenge is split into two scripts:

 * The server 'cp_test_c31_server.py' implements the challenge's server
   side. It should be started before running the client script. It's
   launched by simply executing the script without any parameters. This
   will start the challenge's simple web server that will be ready for
   processing client requests. Once done, the server can be stopped with
   "Ctrl+C".

 * The client 'cp_test_c31.py'. Takes one parameter: the file name used
   for signing on the server side and whose signature the client script
   will break. The server script must be already running when the client
   is started.

The server, as required by the challenge, creates an HTTP server
instance and runs it forever ("Ctrl+C" aborts it), listening to HTTP GET
requests for a specific URL and Query String (specifying a file name and
an associated signature) and does the following:

 * Some initial validations on the GET request (path, query string,
   etc.).

 * Calculates the HMAC-SHA1 associated with the file (its signature).
   In this case, for simplifying, it's just hashing the file name
   itself.

 * Compares the calculated signature against the one given in the URL's
   query string, using the 'insecure_compare()' routine as specified in
   the challenge's statement. It is the 'insecure_compare()' function
   that provides the timing leak that will be exploited by the client
   script to break the signature for any given file name.

 * Provides the HTTP response with the required status ('500' when
   signature is not correct and '200' otherwise).

On the other side, the client will perform the attack described in the
challenge, exploiting the facts that server leaks information about the
correctness of signature bytes via the response timings: each correct
byte delays the request by 50 ms. Additionally, knowing that the server
does a byte-by-byte comparison on the signature, that it responds as
soon as an incorrect byte is found and that it responds with status
codes that allow confirming when a signature is correct, the client
organizes a brute force attack on the signature:

 * Guessing each signature byte sequentially, testing for all possible
   values (0~255) for each byte. This is done in the function
   'execute_break_hmac_sha1()'.

 * Issuing an HTTP request for each guess, using the
   'http_check_file_signature()' function which, besides actually
   sending out the request and collecting the response status (200 for
   signature valid or 500 otherwise), also measures the response delay,
   which is the key information for the attack.

 * The breaking function will know it has guessed another key byte right
   when the response delay is increased by 50 ms (the server delay from
   'insecure_compare()'). This is controlled by 'delay_min' in the
   'execute_break_hmac_sha1()' function.

 * When all signature bytes are guessed, a final request is issued for
   confirming the signature. This final request shall have a response
   with the '200' status.

The client scripts shows progress about the signature breaking, which
can take a while (~30 minutes), depending on the actual signature. Due
to the way the server checks the signature, each byte guess check will
take more 50 ms than the previous one, so the last bytes will take much
more time to guess than the first ones.

As usual, in the end, the client scripts show some useful information
and shows the test success.

New utilities have been added to 'cp_aux_utils' module, required by
this challenge:

 * Function 'sha1_hmac()': Implements the HMAC-SHA1 signing.

 * Classes 'CpHTTPServer' and 'CpHTTPServerReqHandler': Implement a
   simple HTTP Server and a way to register an external function to
   process the HTTP GET requests. Built on top of python's 'http.server'
   and 'urllib.parse' modules.