Reduce verbosity on C31 server
Too many output messages are messing with the server timings for serving
requests. The response timing accuracy is a key factor for the
challenge.

Overridden the HTTP module's 'log_message()' to filter out its output
messages.

Moved some 'print' messages to new 'debug_msg' messages, disabled by
default.

Added a new 'print', only issued when a request for a new, different
file arrives, to have a way of monitoring the requests for each new file
and the associated signature.