Refactor C31 client for reusability
The difference between the upcoming C32 client and the C31 client is
only the delay associated to the server and the actual function used for
breaking the HMAC_SHA1.

This way, to avoid code duplication, refactored C31 client code so that
it can be reused by upcoming C32 client, by:

 * Moving the main script code to a new 'main()' function which gets the
   delay and breaking function by parameter. Also gets by parameters the
   script name and the challenge's title, so that the correct challenge
   info is output on the messages.

 * Changing the main script to simply call the new 'main()' function
   with the C31 parameters.

With this refactoring, the server for C32 can be done by simply calling
C31's 'main()' function with the correct parameters for C32 (its script
name, title and specific delay and breaking function).