Implement Challenge 32: Break HMAC-SHA1 low delay
Added server and client scripts for "Challenge 32: Implement and break
HMAC-SHA1 with a slightly less artificial timing leak".

This challenge is in all similar to C31, except for:

 * Server uses a smaller delay in its 'insecure_compare()' function:
   using 3 ms instead of C31's 50 ms.

 * The client uses an improved breaking function, to deal with C31's
   breaking function flaws when using smaller delays.

This way, except for the different delay and breaking function, C32
server and client scripts are built on top of C31 scripts. The server
simply calls C31's 'main()' function with its own script name,
challenge title and specific delay. The client also simply calls C31's
main function with its own script name, challenge title, specific delay
and own breaking function.

The breaking function in C32 is an improved version of C31's to deal
with C31's flaws when the server delays are smaller.
C32 uses a more statistical approach, by measuring the delays all
possible values for each signature byte, calculating average delays and
guessing the correct byte with information extracted from those
averages and deviations. Additionally, it also does consistency checking
on the delays for the bytes that are (presumably) already guessed and,
when something is not correct, it can go back to previous bytes and
retry the guessing.

Besides the more statistical approach for having a better informed way
of guessing the signature bytes, the principles used are the same as for
C31 (commit 4688fcd).