Refactor folder structure
Reworked the repo's folder structure. All the files in the root folder
was becoming too confuse.

The scripts for running the challenges ('cp_test_c*.py') are now all in
the 'src/' folder. They must be run from this folder.

The data files (input data, output data, dictionary, etc.) are now
located in 'src/data/'. Touched all scripts that use data files to
update them to the new data path.

The utilities ('cp_aux_utils.py') are now in 'src/utils/'. In order to
be able to import from this folder, an empty '__init__.py' file was
added, as per python rules for modules. Touched all scripts that
imported the utilities to update the import statement.