Implement Challenge 34: MITM key-fix attack on DH
Added client, server and MITM (Man in The Middle) attack scripts for
implementing "Challenge 34: Implement a MITM key-fixing attack on
Diffie-Hellman with parameter injection".

The challenge is implemented through 3 scripts:

 * The server ('cp_test_c34_server.py'):
    * Implements the challenge's real server side protocol using a TCP
      server.
    * Must be started before running the client.
    * Takes as arguments the host and port where it listens to requests
      (e.g. 'localhost' and '9001').
    * "Ctrl+C" can be used to abort the server.

 * The MITM attacker ('cp_test_c34_mitm.py'):
    * Implements the MITM server/client protocol described by the
      challenge, i.e., acts as a server to the client and as a client
      from the real server.
    * Performs the parameter injection attack for key-fixing.
    * Must be started before running the client (if the client is
      configured to access the MITM server instead of the real server).
    * Must be started before running the client.
    * When serving the client, the real server must also be already
      running.
    * The arguments are the host and port where the MITM server listens
      to and also the real server's host and port.
    * "Ctrl+C" can be used to stop the MITM server.

 * The client ('cp_test_c34.py'):
    * Implements the challenge's client side protocol.
    * Can be used to connect directly to the real server (e.g., for
      testing the protocol), but in the challenge's context, it connects
      to the MITM server. The host and port used as arguments will
      define to which server the client connects to.
    * The server that it connects to must be running. If the server is
      the MITM server, then the real server must also be running.
    * The arguments are the server (either MITM or real) host and port
      plus the test message (enclose it in '' for dealing with spaces).

Some new utilities classes have been added to 'cp_aux_utils' module,
required by this challenge:

 * Classes 'CpTCPServer' and 'CpTCPServerReqHandler': Implement a
   simple TCP Server and a way to register an external function to
   process the connection requests. Built on top of 'socketserver'
   module. This functionality is used by both the real server and the
   MITM server implementations, in order to provide their server-side
   functionality.

 * Class 'CpSocketIO': For simplifying read/write operations over
   sockets, with respect to the challenge's needs. Implemented on top of
   'socket' module. Used by all scripts for sending and receiving data
   over sockets.

The client uses the 'socket' module to connect to the server configured
by the script's arguments (can either be the MITM or the real server).
Then, it implements the protocol as described in the challenge's
statement and collects all relevant data built in the meanwhile (the
keys being used and the decrypted server echo message which should match
the original message sent to the server). All script arguments are
displayed as well as the results and relevant data.

The real server uses the new TCP Server utility classes to establish a
server listening on the host and port specified by the script's
arguments. Upon new connection requests, it implements the server side
protocol as described in the challenge's statement.

The MITM attacker establishes itself as a "fake" server, using the same
methodology as the real server, on the host and port specified in the
script's arguments. Upon new connection requests from the client, it
implements the MITM protocol described in the challenge's statement.
By injecting 'p' to the real server (instead of client's 'A') and
injecting 'p' to the client (instead of server's 'B'), the MITM attacker
will in fact force that both client and server reach a shared session
key with value of zero, thus being able to actually select the key used
for encryption/decryption, allowing him to also be able to decrypt all
the messages exchanged.

Example usage for the scripts:

 * Use only the client and real server:
    `python cp_test_c34_server.py localhost 9001`
    `python cp_test_c34.py localhost 9001 'Test message for real server'`

 * Use client, real server and MITM attacker:
    `python cp_test_c34_server.py localhost 9000`
    `python cp_test_c34_mitm.py localhost 9000 localhost 9001`
    `python cp_test_c34.py localhost 9001 'Test message for MITM'`

All scripts provide some output on what's going on and the client
script provides, as usual, info about relevant data and overall test
result. There's a 'debug' flag (defaults to 0) on all scripts' code
which can be set to '1' to enable more verbosity and debug messages.

The 'README.md' file has also been updated with new data brought in by
this challenge (new usage and examples).