Don't let Diffie-Hellman secret key be zero
Do not allow zero as a valid value for the secret DH key generated by
the 'dh_keys()' utility function.

The secret key is used as the exponent in modular exponentiation
operations done by DH. Using zero exponent (always results in '1' not
matter what the base is), can mess with other calculations used in the
scripts.