Implement Challenge 36: Implement SRP
Added client and server scripts for implementing "Challenge 36:
Implement Secure Remote Password (SRP)".

The challenge is implemented in 2 scripts:

 * The server ('cp_test_c36_server.py'):
    * Implements the challenge's server side protocol using a TCP
      server.
    * Must be started before running the client.
    * Takes as arguments the host and port where it listens to requests
      (e.g. 'localhost' and '9001') and also the user and password to be
      verified from connecting clients.
    * "Ctrl+C" can be used to abort the server.

 * The client ('cp_test_c36.py'):
    * Implements the challenge's client side protocol.
    * The server script must be running before executing the client.
    * The arguments are the server host and port plus the user and
      password to be validated by the server.

New utilities functions related to SHA-256 and required by this
challenge have been added to the 'cp_aux_utils' module:

 * 'sha256_mac()': Produces a SHA-256 MAC.

 * 'sha256_hmac()': Produces a SHA-256 HMAC.

The new utility functions were implemented on top of python's 'hashlib'
and 'hmac' modules.

The client uses the 'socket' module to connect to the server configured
by the script's arguments. Then, it implements the client-side SRP
protocol as described in the challenge's statement and collects all
relevant data generated and collected in the meanwhile (parameters, keys
and final user/password authentication result message). All script
arguments are displayed as well as the results and relevant data.

The server uses the TCP Server utility classes to establish a server
listening on the host and port specified by the script's arguments. Upon
new connection requests, it implements the server side SRP protocol as
described in the challenge's statement. In the end, it validates the
user/password sent by the client via SRP against the user/password given
to the server via its arguments.

Example usage for the scripts:

 * User/Password "OK":
    `python cp_test_c36_server.py localhost 9001 myuser mypass`
    `python cp_test_c36.py localhost 9001 myuser mypass`

 * User/Password "NOT OK":
    `python cp_test_c36_server.py localhost 9001 myuser mypass`
    `python cp_test_c36.py localhost 9001 myuser mypassword`
    `python cp_test_c36.py localhost 9001 myusername mypass`

All scripts provide some output on what's going on, and the client
script provides, as usual, info about relevant data and overall test
result. There's a 'debug' flag (defaults to 0) on all scripts' code
which can be set to '1' to enable more verbosity and debug messages.

The 'README.md' file was updated with the new python dependencies
('hashlib' and 'hmac') introduced by the new SHA-256 utility functions.