Implement Challenge 37: Break SRP with a zero key
Added client and server scripts for implementing Challenge 37.

This challenge is fully based on C36, so its client and server scripts
are reusing the C36 ones as well.

The main difference in functionality from this challenge to C36 is the
new possibility for force the client's 'A' value to be a multiple (>= 0)
of 'N', which will force the server to set it's 'S' key to zero, no
matter what the multiple of 'N' was forced in client's 'A'.

The client script from C36 (reused in C37) was extended so that its
'main()' function now has a new 'allow_forced_A' parameter indicating if
the user can selected a forced 'A' value or not (set to 'True' in C37
and 'False' in C36). When the value is 'True', the client script usage
allows a new optional argument (the last), named '<forced_A_m>', for
forcing the client's 'A' to be 'N * <forced_A_m>', for making the attack
described in the challenge's statement.

The server script of C37 is fully reused from C36's, without any changes
at all.

Apart from the differences highlighted, C36 and C37 share the same
implementation, protocols and outputs. Refer to C36 for the
implementation details.

Example usage for the scripts:

 * User/Password "OK", without forcing 'A':
    `python cp_test_c37_server.py localhost 9001 myuser mypass`
    `python cp_test_c37.py localhost 9001 myuser mypass`

 * User/Password "NOT OK", without forcing 'A':
    `python cp_test_c37_server.py localhost 9001 myuser mypass`
    `python cp_test_c37.py localhost 9001 myuser mypassword`
    `python cp_test_c37.py localhost 9001 myusername mypass

 * User/Password is always "OK", when  forcing 'A', even with incorrect
   password given:
    `python cp_test_c37_server.py localhost 9001 myuser mypass`
    `python cp_test_c37.py localhost 9001 myuser wrongpass 0`
    `python cp_test_c37.py localhost 9001 myuser wrongpass 1`
    `python cp_test_c37.py localhost 9001 myuser wrongpass 123`